Update patch set 3
Patch Set 3:

This doesn't do what you think it does.

read_delta_qindex() is tied to Cisco's superblock-level quantizer control. It does not control the quantizer on a block level.

It gets called for every block, but on most of them it does nothing. The actual condition to read bits from the bitstream is:

  if ((bsize != BLOCK_LARGEST || mbmi->skip == 0) && (b_col == 0 && b_row ==0))

What this means is that if bsize _is_ the largest supported partition size and it is skipped, then no delta_q value is read. Otherwise, a delta_q value is read only for the the block in the very uper-left corner of the superblock.

In order to actually do what you want, you need to change the level that segment_id's are coded. This should probably be coordinated with Rostislav, who I thought was looking at splitting off the quantizer changes from the other things in the segment map.

Patch-set: 3
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>