General tidy-ups in loop restoration code
The main change here is to change the type of the function that
filters a single loop restoration unit and to move the iteration over
those units to a common function.

There are also changes to extend const-ness for the source
frame (locally broken by the striped restoration code, but I think
that's clearer).

Some of the ugly #if/#endif blocks have been elided in the wiener
filter code (both low and high bit depth), by defining a convolve
alias based on USE_WIENER_HIGH_INTERMEDIATE_PRECISION.

Change-Id: I18043a172ef0ca1154d87cf7f63e3a80944627cd