Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > > Patch Set 7:
> > > 
> > > > I was just looking over this patch again, and thought: What if, instead of setting up a default warp model when the signalled one is invalid, we instead just reset the motion mode to SIMPLE_TRANSLATION?
> > > 
> > > Seems good to me.
> > > 
> > > > You could of course build a model which has the right translational part and go through the warp filter path, but it's a lot more work.
> > > 
> > > Yes, especially for hardware designers. Let's keep it simple :-)
> > 
> > SIMPLE_TRANSLATION is the motion mode used for blocks that use global motion. Motion modes other than SIMPLE_TRANSLATION (ie WARPED_CAUSAL and OBMC_CAUSAL) are only associated with blocks that use obmc or local warp.
> 
> Sorry, I assumed this was relevant to all warped models, not just local warp. Ignore my previous comment. David's suggestion seems ok to me.

Upon looking at the reasoning for the assertion failure mentioned earlier, it seems that it is necessary to set the motion mode to SIMPLE_TRANSLATION even if you decide to use the default parameters. In reconinter.c, the decision to go through the warping functions for a local warped block is determined based on the motion mode.

Patch-set: 7
Reviewer: Gerrit User 5186 <5186@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>