Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > fyi: https://aomedia-review.googlesource.com/c/aom/+/28046
> > > > 
> > > > Not sure which one we should land, but I'd argue the entire adopted experiments magic flag handling needs to go. Brokenness like this wouldn't happen.
> > > 
> > > The following config command proves itself's usefulness and simplicity:
> > > $ configure --enable-experimental --disable-adopted-experiments --enable-xyz
> > 
> > But why do we need to disable adopted experiments at all-- what's the purpose of turning things off that have been accepted as part of the codec?
> 
> This helps to find build warning/error, and some codec issue. Some of them are valid issue. We are also interested in individual experiment's cycle contribution on top of the baseline. This simple command line helps engineer's communication.

I disagree-- in my opinion the main thing it provides is an invitation for trouble. It allows easy creation of a synthetic environment that masks errors and warnings caused by experiment interactions. 

Also, if adopted experiments aren't baseline, what is the baseline? 

Given the question we should probably take this part of the discussion offline. Perhaps it needs to be discussed w/the entire group as this small audience is inadequate to capture everyone's needs, or to provide full understanding of the issue at hand.

Back on topic: one of these patches need to land because the cmake/configure sync checks really need to be enabled sooner rather than later so we can avoid problems caused by unnoticed config mismatches.

Patch-set: 4
CC: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>