Define av1_foreach_rest_unit_in_frame
This is the last stage in a quest to move all knowledge of the layout
of restoration units across the frame into restoration.c. Now this is
done, we can change how they are laid out (to split them properly at
tile boundaries) without having to change code in any other file.

Change-Id: Id5108d787d342f5070580d0e34d84b5ddcc53a86