Update patch set 2
Patch Set 2:

> This CL causes a segmentation fault when encoding (see issue 944 on
 > Monorail)

The reason of seg fault is that the code uses a trick to store parameters in memory that is not aligned (second_pred). So in the certain test case, it fails. To amend this, it requires refactoring of code, which is probably long.

I will have patches for the refactoring, in the future.

Patch-set: 2
Reviewer: Gerrit User 6585 <6585@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>