Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> Hi Peter, would something in these lines do the job?

This certainly looks better than the previous patch as it avoids changing the syntax decode process.

However, at this point in development, it feels like signalling an error at this point would be even better because otherwise encoder bugs may be masked and the performance analysis runs on AWCY become invalid.

The problem is that an encoder should never be choosing an invalid set of warp parameters, so it is good to be able to detect when this occurs.

Patch-set: 4
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>