Update patch set 4
Patch Set 4:

All the encoder has to do is to signal a non-warp motion_mode when invalid parameters are detected (this signalling is always an option and is detected before the warp is applied).

At the moment this change forces both the encoder and the decoder to perform the check (to avoid encoder/decoder mismatch) - which seems like more work than just getting the encoder to do it?

(The transform coefficient problem is a slightly different case because the overflow may only be detected when an encoder has committed to using a particular transform so it is much harder to handle bad cases gracefully.)

Patch-set: 4
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>