Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > AWCY results should eventually appear at https://arewecompressedyet.com/?job=13280-pre&job=13280-post
> > > 
> > > Most of the encodes have finished on AWCY; the summary is that this patch gives about a 0.1% BDRATE improvement
> > 
> > pre seems to be better than post. Is pre the results with this patch?
> 
> Both of those runs seem to have failed, giving incomplete results.

It looks like a handful of the videos failed to decode, even after retrying the encode a few times. I'm not sure why - it would need to be investigated further.



That said, 'pre' is definitely the parent of this patch and 'post' is with this patch.

Probably what happened is that, when I checked originally, the ones which failed were being retried and so weren't counted. Then they seem to have been given large negative BDRATE values (not sure where that's derived from if it failed...), and this dragged the average down from a 0.1% gain to a 0.1% loss.

Patch-set: 5
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>