Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> shouldn't the cmake experiment check have caught this when removed from configure?

It was removed from both in the same patch[1]-- this is just the noise in the deps check that forced restrict_compressed_hdr to on when new_multisymbol was enabled. 

As for why it wasn't caught: Once removed from aom_config_defaults.cmake the variable will no longer show up in aom_config.h. Doesn't matter if the deps check sets it or not when it's not within the defaults because of the way generate_aom_config_templates.cmake is written. 

[1] https://aomedia-review.googlesource.com/c/aom/+/29001

Patch-set: 1
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>