Update patch set 4
Patch Set 4:

> That is a fair point about the encoder complexity - although if the encoder ever did use such an invalid case it would likely be getting very poor compression (because it selected a warp mode, but ended up using translation) so it probably needs to check this in any case.

It might be OK for an encoder to sometimes badly compress a block, if it only happens on rare degenerated situations. In this case, the computing cost of the extra check might not be worth the video quality benefit.

> I'd also be happy with a warning message - although my understanding is that the encoder is currently incorrectly generating invalid cases so I am not sure I would trust an encoder-side warning message to work correctly...

An encoder-side assertion might be trustable, provided it's computed the same way the decoder does.

Patch-set: 4
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>