Update patch set 4
Patch Set 4:

> An encoder-side assertion might be trustable, provided it's computed the same way the decoder does.

But I'm also ok with a warning on the decoder side.

In the meantime, I suggest we get this very CL merged, as it's clearly a net win when it comes to parsing/reconstruction independence.

Patch-set: 4
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>