Adding a speed feature for tx_size search
This patch factors out a function that computes the rd cost for
a given transform type given the transform partition already
computed. This is then used to develop a speed feature where the
transform size search disables trellis optimization but once the
transform sizes are decided, a final search is condcuted with
optimization turned back on.

Change-Id: I30acfc5e2dd353d711e5f4260d5b344847b03ade