Update patch set 6
Patch Set 6:

Hi Sebastien,

I like the idea of moving sortSamples() and possibly the entire sample search until after we know that motion_mode == WARPED_CAUSAL.



With regard to the "fallback" behaviour, Peter and I thought about it more while looking at the current version of the spec. The spec does things very differently - the model generation and validation happen in the equivalent to allow_warp(). The advantage of doing things this way is that the syntax stage doesn't need to care about how the model generation works.

Within this framework, a cleaner solution would be to just have allow_warp() return 0 when it finds an invalid model. That would cause us to use the regular interpolation filter, which can pick up the block's motion vector as normal. We could also do the same thing for global motion - the reference code currently errors out if a global motion model is invalid, which raises the same resilience concerns you had about local warp models.



In terms of the current code, with the model generation being performed during syntax decode (*), you could match the above proposal by marking each model as "valid" or "invalid" when it's generated. Then you can leave the motion mode, generated model, etc. as-is (**). All that needs to happen is for allow_warp() to return 0 for invalid models.

What do you think about that approach?



(*): As I understand it, we can't change where model generation happens in the current code, because the ext-warped-motion experiment uses the local warp models to help predict subsequent blocks' motion vectors. So it requires the model generation to happen during syntax decode.

(**): I'm using "valid"/"invalid" as shorthand for "is/isn't compatible with the warp filter", but even "invalid" models can still be useful - the global motion models are used for motion vector prediction, and that has no restrictions on the model used. As mentioned above, with ext-warped-motion this applies to local warp models too.

Patch-set: 6