Update patch set 1
Patch Set 1:

Introducing a crash is always risky, even if it's in a disabled experiment. 

The experiment is going to be enabled, and the crash might have been forgotten at this point (especially if it's non-trivial to fix).

Moreover, it's a lot harder for us to verify if other changes breaks this very experiment, if it already crashes by default.

Patch-set: 1
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>