[CFL] Only use CfL with the recode loop
Chroma from Luma prediction is based on the reconstructed luma pixels.
These pixels are not available without the recode loop. As such, we must
disable CfL is the recode loop is not being used.

This change does not alter the encoder behavior, it's intended for some
unit test that do not use the recode loop.

Change-Id: I58c7d7cb5b217b14f50016ce7b4d08a3c5d84c24