loop-restoration: Rework self-guided filter
Because we have an (effective) 3-pixel border around each
processing unit, and the local sums in the self-guided filter are
only taken over at most 5x5 regions, we have 1 pixel's worth of
spare border.

We can use this border to greatly simplify the filter: Instead
of calculating a 64x64 region of the A[] and B[] arrays, we can
calculate a 66x66 region. Then we don't have to deal with complicated
boundary conditions when generating the final 64x64 output block.

In addition, the 'boxnum' function has been effectively redundant
for a while - due to the way we do the 5x5 (or 3x3) windowing,
the values we actually use are always (2r+1)^2. So we can skip
the entire function, as long as the maximum r is <= 2.
But keep it available in case we ever want to use r > 2.

Change-Id: I99beee164a31349a5228a9bef048e5f35c9639f2