Fix self-guided loop restoration filter for tiny processing units
We don't actually apply the self-guided filter to tiny processing units,
since we need a certain number of pixels for the edges to be correctly
filtered.

When skipping these tiny units, the previous code didn't copy the
source across to the destination so just left garbage there. This
patch re-evaluates how small the units may be. The only cases we can't
handle are when the width or height are < 2px, and the patch just
copies data across when it encounters such units.

Change-Id: I0a588f5f71262487db1395d26fbea6f076d89eb8