[CFL] Reorder alpha search for early termination
This prepares the way for a heuristic to truncate the search.

For a given joint sign value, each plane's alpha is coded
independently and so the alpha choice is also independent.
Rather than iterating over all possible pairs, first choose
for each plane separately and then merge the choices.
Merge with RD collection and continuously update the bound.

On subset1, then encoder output is identical for most sequences:
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.0002 | -0.0058 | -0.0016 |   0.0020 | 0.0025 |  0.0040 |    -0.0022

Change-Id: Ie7d1e1622d46eaa8cf42f50c9a6bec6e59bdbc39