[CFL] separate independent parts of RDO search
For a given joint sign value, each plane's alpha is coded
independently and so the alpha choice is also independent.
Rather than iterating over all possible pairs, first choose
for the U plane and fix that value while searching V.

This causes no change in encoder output and the difference
in encoding time is negligible.

Change-Id: Ie7d1e1622d46eaa8cf42f50c9a6bec6e59bdbc39