[level map] simplified context derivation
This CL simplifies context derivation for nz and base level flags in
level map.
1. Reduce SIG_COEF_CONTEXTS from 58 to 42.
2. NZ and base level flags share the same context offsets derived from a
template of size 5 (down from 7 for NZ and 11 for base level flags).

In limited runs, compression performance seems neutral if not better.
Encoding time for a key frame on a local linux machine is reduced by about 25% or more.

Change-Id: Ibd93b21c839154bc5ae26b993f9e66537cbf5942