Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Tom, please forget what I said about gcc ignoring "fallthrough intended".
> > > 
> > > I just realized I got mislead by ccache (which is always on on my PC).
> > > 
> > > gcc emitting different warnings depending on the comment defeats ccache's cache, as now the output of gcc depends on something ccache doesn't see.
> > > 
> > > I have to re-test ; anyway, I think we'd better disable this warning once and for all.
> > > 
> > > James, what's your opinion?
> > 
> > if it can be annotated then leaving it on would be best in my opinion; there is some [[ clang/gcc::fallthrough ]] I think. these could then become a define
> 
> I didn't really want to go down this path, but I agree that it's probably the correct choice in the end. Here are the two flavors.
> 
> GCC:
> __attribute__ ((fallthrough));
> 
> clang:
> [[clang::fallthrough]];

Sorry to not be clear: I don't mean handle the above in this patch.

And, because I'm not clear on this point: Is "// fallthrough intended" working for this particular case?

If it is I think this patch should use the fallthrough intended comment, and then the compiler specific handling can be done later on.

Patch-set: 1
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>