Add Daala TX fixed-coeff-depth capability to quantization
This patch completes the work to add fixed-depth TX domain support to
the quantization and dequantization code.  At present, it is active but
configured to behave identically to current AV1 master as RDO and TX
have not yet been updated to also support this functionality.

subset-1:
monty-rest-of-stack-noshift-s1@2017-11-13T14:37:42.541Z ->
 monty-rest-of-stack-quant-s1@2017-11-13T14:38:43.774Z

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.0000 |  0.0000 |  0.0000 |   0.0000 | 0.0000 |  0.0000 |     0.0000

objective-1-fast --limit=4:
monty-rest-of-stack-noshift-o1f4@2017-11-13T14:37:16.992Z ->
 monty-rest-of-stack-quant-o1f4@2017-11-13T14:38:28.828Z

  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.0000 |  0.0000 |  0.0000 |   0.0000 | 0.0000 |  0.0000 |     0.0000

Change-Id: I3773a1fc128136c9fea227f4b547576a8aa6efa3