Speed up of ext-partition types
Search the new horz/vert a/b/4 partitions only if the best so far
is either oriented along the same direction or split/none, or if
the rd costs obtained from the previous partition searches indicate
there is potential in searching these partitions.

The pruned search is implemented in two levels of a speed feature.

This brings about 25-30% speedup.

Change-Id: I6c6c347e06c34ee0ca17479aeeb4075a66dc7e2c