Speed up of ext-partition types
Search the new horz/vert a/b/4 partitions only if the best so far
is either oriented along the same direction or split/none, or if
the rd costs obtained from the previous partition searches indicate
there is potential in searching these partitions.

This brings about 25-30% speedup at less than 0.1% drop as seen on
lowres 30 frames.

Change-Id: I6c6c347e06c34ee0ca17479aeeb4075a66dc7e2c