Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> The experiment tracker now has a link to the design document.
> 
> max_frame_{width,height} are BOTH default and maximum values.
> 
> The approved proposal expresses the limitation on frame size in the semantics section.

Okay, thanks for the clarification, and thanks for adding the link on the experiment tracker :)

> Personally I have no objection to using aom_read_primitive_quniform, just in doubt how much can be changed without WG approval, especially, since aom_read_primitive_quniform seems to be a "variable-length code" which would be the only place in the frame header I believe.

Okay, that seems fair enough. Note that we do already have some variable-length syntax elements in the form of the global motion parameters.

Another solution would just be to add a check that the signalled frame size is at most the maximum size, and call aom_internal_error() if that doesn't hold.

Patch-set: 5