Misc. clean ups / refactor of speed 1
With this patch, and the speed settings turned on for speed 1,
the coding efficiency of speed 1 in default configuration should be
only a little worse than speed 0, but it should roughly run at
double the speed.

Specifically, this patch makes various changes to make sure that
speed 1 behaves exactly the same as speed 0 except for speed settings
turned on or off in speed_features.c.

This will change the bitstream generated a little for speeds
1 or higher because of the following reasons:
1. Removes a hacky speed setting correction factor in firstpass.c
2. Fast cdef search is moved from speed 1+ to 2+, and a new speed
feature is added to control that.
3. Mesh search settings are pushed down one level so that speeds 0
and 1 use the same settings.
4. A disable_split_mask feature for animated content previously
turned on speeds 1+ is moved down to speeds 2+.

Change-Id: I0ec36556f157bdc42c5daa0cfb9518cf7ff65f6b