Speed up ref frame encoding using GF structure
This patch is to speed up the encoder by pruning the forward
references, following rules of:

When both forward reference and backward reference are available,
check whether all the forward refernce frames are from the current
GF group. If not, keep (1) GOLDEN frame and (2) the latest forward
reference frame, and disable all the other forward references from
the previous GF group.

Change-Id: Ie806a00a33920633ac914d1cb7b6f636ad72fac2