Fix mismatch with striped loop restoration+superres
When upscaling a frame, we extend frame borders to stop the upscale to
save boundary lines convolving with uninitialised data off the edges,
which was causing encode/decode mismatches. With this patch, we only
do the extension when there's going to be an upscale (otherwise
there's no need), which should give a small coding gain when not
upscaling.

More importantly, it forces us to extend in the decode path whether or
not we are using loop restoration, which matches what the encoder does
and fixes a mismatch.

Change-Id: Ie5a0791b0cbedbf254f9080f3cbf668318673f2f