Remove ONE_SIDED_COMPOUND experiment
Always assume ONE_SIDED_COMPOUND is true as it has been adopted
officially. Further, this patch also fixes the cmake setup for
the experiment of EXT_COMP_REFS, as when EXT_COMP_REFS is on, it
shall always have ONE_SIDED_COMPOUND on (but that was not the case
without this patch).

We leave EXT_COMP_REFS for the time being, even though it is also
adopted, since by keeping this flag, it may allow its turning off to
tune the encoder speed. The experiment of EXT_COMP_REFS should be
removed eventually at the closing of the bitstream.

Change-Id: I0644318e522f5b48d5bba4e4391104a24d43cdc3