CFL: Fix validation code
When performing a CfL prediction where we need to take pixels
from multiple luma blocks, there is some code to check that
we've correctly filled in 'cfl->pred_buf_q3' before attempting
a prediction. This is designed to catch cases where we use
stale data, while still allowing the encoder to try out multiple
predictions for the same block.

This validation code was buggy in the case of inter blocks -
while we'd store the luma pixels just fine, we would only mark
the last (luma) transform unit within each inter block as "filled in".
This accidentally worked for 4x8 and 8x4 blocks using 4x4 transforms,
as the top-left 4x4 unit is always marked as "filled in", no matter what.
But 4x16 and 16x4 blocks didn't validate properly.

Since we're actually filling the array correctly, we just need
to change the validation code to mark correct region as being filled in.

Change-Id: I24bde52bfb49195d6829baca60ec7b41db853aa0