Add experimental flag of SCALE_REFMV
This patch is to add a scaled version of the mv(s) of the spatial
neighboring block, if any and if the reference frame of the
neighboring block is different from the current reference frame,
to the reference motion vector list. The scaling is done according
to the temporal frame distance.

Change-Id: I40255355978825dc0c056cb184082d3196326501