Add experimental flag of SCALE_REFMV
When getting the motion vector reference candidates from the spatial
neighboring blocks, if the reference frame of the neighboring block
is different from that of the current frame, this patch will add a
scaled version of the mv(s) of the spatial neighbors, according to
the temporal distance between frames, instead of adding the mv as is
excepting flipping the signs if needed.

Without the use of mfmv, this patch obtains a coding gain at level of
~-0.25% in BDRate saving; With the use of mfmv, it obtains ~0.1% in
bitrate saving.

For instance, the biggest gain of this patch has been obtained in the
clip of Keiba:
Keiba_240p:    (vs w/o mfmv) -1.906% (vs with mfmv) -0.899%
Keiba_832x480: (vs w/o mfmv) -2.556% (vs with mfmv) -1.042%

Change-Id: I40255355978825dc0c056cb184082d3196326501