Update patch set 9
Patch Set 9:

This new version of the patch fixes the failing assertion (which was caused because I didn't think about the fact that av1_get_tile_rect clips to the frame boundary, when we want to think in whole blocks of 8 luma pixels here).

I've removed the double assertion again, and rewritten it in the form of "The last row we need fits in the frame", which is less confusing than it was originally and makes it a bit more obvious what's going on.

I've also just realised that I accidentally removed the "// NDEBUG" on the #endif line - I'll push up a new patch in a sec putting that back in.

Patch-set: 9