Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> LGTM.
> 
> Do we have any performance change numbers?

I indeed tried another patch, which was more aggressive to try out on the mv projection in mfmv considering the different options for cur_to_bwd, etc.:
https://aomedia-review.googlesource.com/c/aom/+/32701
But I did not observe any noticeable coding performance change of mfmv2 as opposed to mfmv for the above patch. I guess for such cases, these mv candidates would not be sufficiently effective to be considered any way.

For the mv projection operation itself, I believe it may be better to have it handle the signed rounding, such that we do not have to consider the sign bias explicitly but to let the projection itself to handle the mv direction automatically.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>