Update patch set 34
Patch Set 34:

Hi Zoe, thanks for reviewing!

> I think you may do a +2 yourself next time as you got one already and the patch does not have to rely on another one just due to rebase.

When I get one +2 on a CL, this +2 is only conserved through trivial rebases (which are only possible if there are no conflicts).

Typically, a CL removing an experiment is a conflict magnet ; so if the CL isn't quickly merged, conflicts occurs and the CL needs to be non-trivially rebased. This loses the +2, and only someone with +2 permissions can put it back (typically, people from Google, Cisco or Mozilla).

This is my current understanding ; however, as I only learnt those rules empirically, I might have missed something ; please let me know if this is the case!

Patch-set: 34